vsphere: use packer to create the podvm template
Mostly, derived from its libvirt counterpart, this script creates
an esx guest from a standard ubuntu iso and then converts it into a template.

Standard settings are in settings.auto.pkrvars.hcl. esx config is
expected in esx.auto.pkrvals.hcl which will be created by the Makefile
if not present; user-data.pkrtpl.hcl contains the autoinstall template.

The automated input of characters at boot to start autoinstall is kind of
flaky, the currently working sequence is defined in boot_command of the main
script.

Signed-off-by: Bandan Das <bsd@redhat.com>